pyproject.toml
[project]
name = "rag-mongovectorstore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "pymongo>=4.15.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
